# book-network-ui/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
          "title": "OpenApi specification - Yusuf",
          "description": "OpenApi documentation for Spring security",
          "termsOfService": "Terms of service",
          "contact": {
                "name": "Yusuf",
                "email": "yusufyuksek513@gmail.com"
          },
          "license": {
                "name": "Licence name",
                "url": "https://some-url.com"
          },
          "version": "1.0"
    },
    "servers": [
          {
                "url": "http://localhost:8088/api/v1",
                "description": "Local ENV"
          },
          {
                "url": "https://some-url.com",
                "description": "Prod ENV"
          }
    ],
    "security": [
          {
                "bearerAuth": [ ]
          }
    ],
    "paths": {
...
          "/auth/register": {
                "post": {
                      "tags": [
                            "Authentication"
                      ],
                      "operationId": "register",
                      "requestBody": {
                            "content": {
                                  "application/json": {
                                        "schema": {
                                              "$ref": "#/components/schemas/RegistrationRequest"
                                        }
                                  }
                            },
                            "required": true
                      },
                      "responses": {
                            "202": {
                                  "description": "Accepted",
                                  "content": {
                                        "application/json": {
                                              "schema": {
                                                    "type": "object"
                                              }
                                        }
                                  }
                            }
                      }
                }
          },
          "/auth/authenticate": {
                "post": {
                      "tags": [
                            "Authentication"
                      ],
                      "operationId": "authenticate",
                      "requestBody": {
                            "content": {
                                  "application/json": {
                                        "schema": {
                                              "$ref": "#/components/schemas/AuthenticationRequest"
                                        }
                                  }
                            },
                            "required": true
                      },
                      "responses": {
                            "200": {
                                  "description": "OK",
                                  "content": {
                                        "application/json": {
                                              "schema": {
                                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                              }
                                        }
                                  }
                            }
                      }
                }
          },
...
          "/auth/activate-account": {
                "get": {
                      "tags": [
                            "Authentication"
                      ],
                      "operationId": "confirm",
                      "parameters": [
                            {
                                  "name": "token",
                                  "in": "query",
                                  "required": true,
                                  "schema": {
                                        "type": "string"
                                  }
                            }
                      ],
                      "responses": {
                            "200": {
                                  "description": "OK"
                            }
                      }
                }
          }
    },
    "components": {
          "schemas": {
...
                "RegistrationRequest": {
                      "required": [
                            "email",
                            "firstname",
                            "lastname",
                            "password"
                      ],
                      "type": "object",
                      "properties": {
                            "firstname": {
                                  "type": "string"
                            },
                            "lastname": {
                                  "type": "string"
                            },
                            "email": {
                                  "type": "string"
                            },
                            "password": {
                                  "maxLength": 20,
                                  "minLength": 6,
                                  "type": "string"
                            }
                      }
                },
                "AuthenticationRequest": {
                      "required": [
                            "email",
                            "password"
                      ],
                      "type": "object",
                      "properties": {
                            "email": {
                                  "type": "string"
                            },
                            "password": {
                                  "maxLength": 20,
                                  "minLength": 6,
                                  "type": "string"
                            }
                      }
                },
                "AuthenticationResponse": {
                      "type": "object",
                      "properties": {
                            "token": {
                                  "type": "string"
                            }
                      }
                },
...
          },
          "securitySchemes": {
                "bearerAuth": {
                      "type": "http",
                      "description": "JWT auth description",
                      "in": "header",
                      "scheme": "bearer",
                      "bearerFormat": "JWT"
                }
          }
    }
}